Fixing sockets - 2

+++ data/wallets.py
@@ -13,4 +13,4 @@ class Wallet(SqlAlchemyBase):
     session_id = sqlalchemy.Column(sqlalchemy.String, sqlalchemy.ForeignKey("sessions.id"))
     user_id = sqlalchemy.Column(sqlalchemy.Integer, sqlalchemy.ForeignKey("groups.id"),
                                 primary_key=True)
-    money = sqlalchemy.Column(sqlalchemy.Integer)
+    money = sqlalchemy.Column(sqlalchemy.REAL)
